djangoBackend/requirements.txt
django
celery
python-dotenv
djangorestframework
django-cors-headers
django-extensions
djangorestframework-simplejwt
redis
django-allauth
dj-rest-auth
requests
cryptography
pandas
chromadb
PyPDF2
langchain-experimental
langchain-huggingface
python-docx
selenium
beautifulsoup4
pymysql
firebirdsql
cx_Oracle
cockroachdb
psycopg2-binary
pyodbc
django-redis
